Save timetable to device. Use GEN[date] code for generated timetables.
