+++ app/src/main/java/com/github/hwutimetable/CurrentDateProvider.kt
@@ -1,7 +1,9 @@
 package com.github.hwutimetable
 
 import org.joda.time.LocalDate
+import org.joda.time.LocalDateTime
 
 interface CurrentDateProvider {
     fun getCurrentDate(): LocalDate
+    fun getCurrentDateTime(): LocalDateTime
 }
